chore: test all file extensions
