+++ test/integration/hashbang/src/cases/cjs.cjs
