+++ test/integration/hashbang/src/cases/js.js
