+++ test/integration/hashbang/src/cases/mjs.mjs
